feat(capacitor): perform Ionic for copy and sync
resolves https://github.com/ionic-team/ionic-cli/issues/3783
