+++ packages/@ionic/cli/src/commands/capacitor/update.ts
@@ -11,8 +11,8 @@ export class UpdateCommand extends CapacitorCommand implements CommandPreRun {
       summary: 'Update Capacitor native platforms, install Capacitor/Cordova plugins',
       description: `
 ${input('ionic capacitor update')} will do the following:
-- Update each Capacitor native project, such as any dependencies that need updating.
-- Install any discovered Capacitor or Cordova plugins.
+- Update Capacitor native platform(s) and dependencies
+- Install any discovered Capacitor or Cordova plugins
       `,
       inputs: [
         {
